requirements.txt
openai-whisper
moviepy==1.0.3